Revert "Avoids unnecessary diffs when the vgw_telemetry is changed outside the terraform workflow"
This reverts commit ddfdefbd553e67b18318c51bacc486d86416dc85.

Terraform warns about exactly this change, trying to ignore changes
on a computed value. Apparently, this is not the cause of the persistent
diff, as originally thought. Instead, what I believe is happening,
is that when the user *outputs* the entire vpn_connection object,
it is the *output* that is changing. The user can address this by
updating the output so it does not include the "last_status_change"
attribute. There may be a better way of doing this, but this is what
I came up with in a pinch:

    value = merge(
      { for key, value in module.vpn_connection.vpn_connection : key => value if key != "vgw_telemetry" },
      { vgw_telemetry = toset([
        for item in module.vpn_connection.vpn_connection.vgw_telemetry : { for key, value in item : key => value if key != "last_status_change" }
      ]) },
    )